+++ src/driver/DriverOptions.ts
@@ -58,15 +58,6 @@ export interface DriverOptions {
     storage?: string;
 
     /**
-     * Indicates if connection pooling should be used or not.
-     * Be default it is enabled if its supported by a platform.
-     * Set to false to disable it.
-     *
-     * @todo: rename to disablePool? What about mongodb pool?
-     */
-    usePool?: boolean;
-
-    /**
      * Extra connection options to be passed to the underlying driver.
      */
     extra?: any;